mod: refactored dispatchMenu
